+++ services/web/src/components/index.js
@@ -3,5 +3,3 @@ export { default as Breadcrumbs } from './Breadcrumbs';
 export { default as Footer } from './Footer';
 export { default as HelpTip } from './HelpTip';
 export { default as PageLoader } from './PageLoader';
-export { default as CodeBlock } from './CodeBlock';
-export { default as RequestBlock } from './RequestBlock';
