+++ services/web/src/screens/Docs/RequestBlock/templates/curl.js
